TargetingTower based restructuring
