+++ Assets/Scripts/Entity/TargetingTower.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 872b700d500a5e54c83c516701fa4c83
