+++ Assets/Scripts/Entity/Tower.cs
@@ -13,26 +13,12 @@ public abstract class Tower : Entity
 {
     protected ComponentModule module;
     protected bool idle = true;
-    private Coroutine targeting = null!;
     private Coroutine acting = null!;
 
     protected override bool Invulnerable() => Hiding;
 
     [SerializeField]
     protected SpriteRenderer sprite;
-
-    /// <summary>
-    /// Time between (re)targeting attempts
-    /// </summary>
-    [SerializeField]
-    protected float retargetDelaySeconds = 10.0f;
-
-    /// <summary>
-    /// Time it takes to choose a target
-    /// </summary>
-    [SerializeField]
-    protected float targetTimeSeconds = 2.0f;
-
 
     [SerializeField]
     protected float actiondelaySeconds = 2.0f;
@@ -55,7 +41,6 @@ public abstract class Tower : Entity
         transform.parent = TowerManager.instance.transform;
         TowerManager.instance.AddTower(this);
 
-        targeting = StartCoroutine(Targeting());
         acting = StartCoroutine(Acting());
     }
 
@@ -89,53 +74,32 @@ public abstract class Tower : Entity
         Hiding = true;
     }
 
-    private void SafeStopCoroutine(Coroutine cr)
+    protected void SafeStopCoroutine(Coroutine cr)
     {
         if (cr != null)
             StopCoroutine(cr);
     }
 
-    private void OnDestruction()
+    protected virtual void OnDestruction()
     {
-        SafeStopCoroutine(targeting);
         SafeStopCoroutine(acting);
     }
 
 
     // call when wave ends / before wave begins
-    public void OnRepair()
+    public virtual void OnRepair()
     {
         HitPoints = MaxHitPoints;
 
         // avoid duplicate coroutines
-        SafeStopCoroutine(targeting);
+        
         SafeStopCoroutine(acting);
 
-        // maybe TODO: StopCoroutines function
-
         // start for real
-        targeting = StartCoroutine(Targeting());
+        
         acting = StartCoroutine(Acting());
 
         idle = true;
-    }
-
-    private IEnumerator Targeting()
-    {
-        while(true)
-        {
-            if (idle)
-            {
-                idle = false;
-                yield return new WaitForSeconds(targetTimeSeconds);
-                //Debug.Log($"[TOWER] : TARGETIMG");
-                Target();
-                idle = true;
-            } 
-
-            yield return new WaitForSeconds(retargetDelaySeconds);
-            yield return new WaitUntil(() => !Hiding);
-        }
     }
 
     private IEnumerator Acting()
